source/depth/pyproject.toml
[project]
name = "depth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.4",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
